Performace benchmarking test for s3select function
Performance tests for s3select functions
